+++ src/components/forms/identities/add-identity-form.tsx
@@ -134,7 +134,6 @@ const AddIdentityForm: React.FC<{
         <InputCheckbox
           name="primary"
           onChange={() => {
-            console.log("Hello");
             setIdentity({
               ...identity,
               primary: !identity.primary,
