+++ tests/e2e/utils/check-verification-code.ts
@@ -8,29 +8,31 @@ const localizedStrings = {
 };
 
 async function getCode(emailAddress: string): Promise<string> {
-  const response = await fetch(
-    "https://mocks.prod.connect.connect.aws.eu-west-2.k8s.fewlines.net/mail/email",
-  );
+  const emailMockUrl =
+    "https://mocks.prod.connect.connect.aws.eu-west-2.k8s.fewlines.net/mail/email";
+
+  const response = await fetch(emailMockUrl);
+
   const emails = (await response.json()) as unknown as {
     headers: { to: string };
     id: string;
     text: string;
     time: number;
   }[];
+
   const email = emails
     .sort((a, b) => (a.time < b.time ? 1 : -1))
     .find((email) => email.headers.to === emailAddress);
+
   if (email) {
     const match = email.text.match(/code (\d{6}) /);
     if (match) {
       const code = match[1];
-      fetch(
-        `https://mocks.prod.connect.connect.aws.eu-west-2.k8s.fewlines.net/mail/email/${email.id}`,
-        { method: "DELETE" },
-      );
+      fetch(`${emailMockUrl}/${email.id}`, { method: "DELETE" });
       return code;
     }
   }
+
   return "";
 }
 
@@ -38,6 +40,7 @@ async function checkVerificationCode(): Promise<void> {
   const shouldDoCodeVerification = await text(
     localizedStrings.sudo.send,
   ).exists();
+
   if (shouldDoCodeVerification) {
     await waitFor(2000);
     await click(localizedStrings.sudo.send);
